(#37) Add CustomRawUrlProvider to enable arbitrary content URLs
A new CustomRawUrlProvider is added that acts as the final IProvider
candidate. This provider interprets the TargetUrl parameter (`-u`) as
the url to use for the RawGitUrl property of the provider.

This enables supporting custom github/bitbucket content proxies as well
as opens up an easy way to support conforming alternative git services
without having to write a new IProvider implementation for each one.