Add support for Visual Studio Team Services #53
Modified SrcSrv.cs to write a different form of data to the pdbs
when vstsData is passed to the function. This form includes using
the SRCSRVCMD with "tf.exe git" to pull the source from Visual
Studio Team Services. So when the Visual Studio Team Services
provider is used, the project name and repository are extracted
from the git url. This information is passed to the CreateSrcSrv
function.