Handle files with a backtick in the name
GitLink uses the `Uri` APIs in order to calculate relative file paths.
This API will escape a number of characters, including backtick, when
converting from `string` to `Uri`. These characters are not unescaped
when converting from `Uri` to `string`. Hence the paths produced from
GetRelativePath are incorrect and GitLink can't process the files.

Fixed to unescape the characters once the relative path is calculated.

closes #163